feat: refresh about page content and share source with resume page
Rewrite the About Me copy and add a Selected work list linking to
shipped changelog entries, and set the current role end date.

Extract the shared About prose, jobs, and selected-work into
src/content/about.tsx so the about and resume pages render from one
source, fixing the previously stale resume/CV.
